Remove CTDB crm resource hierarchy assumption
For clustered AD domain membership, it is incorrectly assumed that the
CTDB base primitive is nested directly under a clone resource.
With this assumption, a configuration specifying the CTDB primitive
under a group resource container would result in failure; (bnc#813462).

Instead, use the CTDB base primitive name when starting or stopping the
resource. crm is smart enough to apply such requests to any parent clone
or group containers.