Allow 'client ipc signing' param be included in generated smb.conf (bsc#1035099)
When yast joins a domain it creates a minimal config file. This config
file though should have more (if not all) of the configuration that is
specified in the 'real' smb.conf file with the exception of any paramater(s)
changed by the user in as a result of interacting with the dialog itself.
As a first step (and as a solution to the bsc#1035099) we will update the
generated configuration with if it exists in /etc/samba/smb.conf

Signed-off-by: Noel Power <noel.power@suse.com>