Fix reading the permitted group to create "user shares"
The ReadSharesSetting function tries to find the permitted group's gid in
the groups listed by 'getent group'. Since samba 4.2.0 winbind does not
enumerate the domain groups, so if the permitted group is a domain group
it is not found and the field is left empty, falling back to the default
"users" groups when saving the configuration. To fix this, Call "getent
group" with the gid number as argument instead parsing the output.

Signed-off-by: Samuel Cabrero <scabrero@suse.de>
(cherry picked from commit 5d8cea11cf961ed316a518a35180b750651203bf)