Fix detection that all IP addresses on a node are valid
Update the detection that all IP addresses on a cluster node are
prepared for CTDB when the system uses DHCP. The code is changed to use
'ip addr' instead of ifconfig to obtain all IP addresses on the node.
The heuristics is also made tighter to check only interfaces that are
really configured with DHCP.

Note that only the first IP address assigned to each interface is
checked for presence in /etc/ctdb/nodes as additional addresses could be
virtual IPs that are managed by Pacemaker+IPaddr2. This was kind of
implicit when using ifconfig because of its limitation to not being able
to display all IP addresses that an interface has.