Avoid use of the removed ctdb --socket option
Release of Samba 4.9.0 (2018-09-13) made the ctdb socket path hardcoded
in ctdb and ctdbd at their build time and the ctdb --socket option got
removed.

Reflect this change in PrepareCTDB() by removing use of the --socket
option. The code can rely on ctdb and ctdbd using the same built-in
socket path.